File size check of the file
I noticed that a lot of times an xml file is saved as a consecuence of an error in the response of the server. These files are small (950 bytes), so i added a check of the .zip filesize to avoid this kind of errors. Also tried with iszipfile routine but got some errors, so i discard that method.

Also i added a counter to know the number of attempts per file.

Hope you like these canges.