Now truly supporting different directories for downloads and optionally overwrite KML/manifest files.
Note that Directory cofiguration item understands also environments
variables so one can use $PWD or $SOME_THING_ELSE to avoid polluting
with paths the configuration file.

Note that it required a change in Sqlite schema to work. It is a bit
redundant, but it works and avoids the use of a paths table.

One can use only one directory per polygon which seems acceptable.